Fix sharing state on clone
We shallow copy steps and mockHandlerCache in clone so that clones get
the same state at the time of clone, but are independent aterwards.

individualExecution was previously relying on clone to apply per
execution mocks (and relied on state being shared). Now that clone
intentionally doesn't share state, just use the constructor directly to
create a new instance.