URGENT: Upgrade required for React CVE-2025-55182 (#1196)
This repository is listed on [Vercel
Templates](https://vercel.com/templates).

It uses a version of Next.js that is vulnerable to
[React2Shell](https://vercel.com/kb/bulletin/react2shell). This PR
upgrades the template to the nearest patched version.

We have temporarily disabled the ability for developers to deploy this
template.

**Action required**

Please review the changes and run a quick test. If everything looks
correct, you can merge this PR.
If you prefer to upgrade manually, feel free to close this and apply
your own fix.

Thank you.

Co-authored-by: Chris Tate <ctate@users.noreply.github.com>